linting changes. renamed the exception
