+++ eva/io_descriptors/eva_data_types.py
@@ -1,10 +1,22 @@
+# coding=utf-8
+# Copyright 2018-2022 EVA
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 import numpy as np
-import torch
 import pandas as pd
+import torch
 
 from eva.io_descriptors.eva_arguments import EvaArgument
-from eva.utils.errors import TypeException
-
 
 
 class NumpyArray(EvaArgument):
@@ -69,7 +81,6 @@ class PyTorchTensor(EvaArgument):
 
         elif not self.dtype:
             return isinstance(input_object, torch.Tensor)
-
 
     def check_shape(self, input_object) -> bool:
         if self.shape:
