+++ eva/utils/errors.py
@@ -16,7 +16,7 @@ class CatalogError(Exception):
     pass
 
 
-class TypeException(Exception):
+class UDFInputOutputTypeException(Exception):
     def __init__(self, message):
         # Call the base class constructor with the parameters it needs
         super().__init__(message)
