+++ test/integration_tests/test_pytorch.py
@@ -305,6 +305,3 @@ class PytorchTest(unittest.TestCase):
         actual_batch = execute_query_fetch_all(select_query)
 
         self.assertEqual(len(actual_batch), 60)
-
-
-
